+++ admin/server/middleware/initList.js
@@ -8,5 +8,6 @@ module.exports = function initList (req, res, next) {
 		req.flash('error', 'List ' + req.params.list + ' could not be found.');
 		return res.redirect('/' + keystone.get('admin path'));
 	}
+
 	next();
 };