+++ admin/server/middleware/request.js
@@ -1,6 +1,6 @@
 const fs 	= require('fs');
 const _ 	= require('lodash');
-const localization = require('../../../lib/handler/localization');
+const localization = require('../../../lib/handler/staticGenerator/localization');
 
 const getStaticLanguageFile = async (nextNode) => {
 	const path = `${nextNode.get('app root')}/${nextNode.get('static lang path')}`;
@@ -14,6 +14,22 @@ const getStaticLanguageFile = async (nextNode) => {
 		const { data: dbLang} = await langHandler.exportLanguageStatic();
 		data = dbLang;
 	}
+	return data;
+};
+
+const getStaticAppLanguageSectionFile = async (nextNode) => {
+	const path = `${nextNode.get('app root')}/${nextNode.get('static section path')}`;
+	var data;
+	try {
+		data = JSON.parse(fs.readFileSync(path, 'utf8'));
+	} catch (err) { // if the language cannnot be read, then query db
+		console.log('> Cannot red the Static App Language Section File, query app language sections from Database.');
+		const appLangHandler = new localization(nextNode, nextNode.list('App_Language').model);
+		// export file, and get db languages, if error then ignore localization in the app
+		const { data: dbLang} = await appLangHandler.exportSectionStatic();
+		data = dbLang;
+	}
+	// console.log(data);
 	return data;
 };
 
@@ -71,5 +87,6 @@ module.exports = async function (req, res, next, nextNode) {
 			langf: frontendCookie,
 		};
 	}
+	req.appLanguage = await getStaticAppLanguageSectionFile(nextNode);
 	next();
 };